Fixed fitness organism function
Fitness organism was not subtracting surplus gain from raw gain,
while the truth organism was. This was causing the fitness organism
to be more fit than the truth organism.

I have fixed this by subtracting the surplus gain from the raw gain
in the fitness organism function. See python and js codes, specifically
the fitness organism function in both, in the fitness organism class.